add benchmark.js benchmark (for #20)
ensures statistical significance
