+++ package.json
@@ -14,6 +14,7 @@
   "license": "MIT",
   "dependencies": {},
   "devDependencies": {
+    "benchmark": "^2.1.0",
     "standard": "^6.0.5",
     "tape": "4.x"
   }
